Linked contact page to /contact url, added examples pages
